Implement Add expense,category,savings

+++ WebApplication1/Models/SavingViewModel.cs
@@ -3,7 +3,7 @@
     public class SavingViewModel
     {
         public DateTime? SavingDate { get; set; }
-        public decimal Amount { get; set; } // holds Amount
+        public long Amount { get; set; } // holds Amount
         public string? SavingNote { get; set; }
 
     }
